moving proptypes outside web components
