+++ packages/@coorpacademy-app-review/.mtslconfig.json
@@ -1,0 +1,1 @@
+{"ignore_dirs":["node_modules"]}
