+++ packages/@coorpacademy-app-review/src/index.test.tsx
@@ -61,6 +61,9 @@ const clickAllSlides = async (
 
 const appOptions: AppOptions = {
   token: process.env.API_TEST_TOKEN || '',
+  templateContext: {
+    plop: 'plip'
+  },
   slide
 };
