+++ packages/@coorpacademy-components/.gitignore
@@ -1,0 +1,2 @@
+# may be generated by mobile app on local dev using mtsl
+.mtslconfig.json
