+++ packages/@coorpacademy-components/src/atom/lottie-wrapper/index.js
@@ -1,5 +1,4 @@
 import React, {useMemo, useRef, useEffect, useState} from 'react';
-import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import lottie from 'lottie-web';
 import get from 'lodash/fp/get';
@@ -9,13 +8,7 @@ import keys from 'lodash/fp/keys';
 import omit from 'lodash/fp/omit';
 import unfetch from 'isomorphic-unfetch';
 import style from './style.css';
-
-export const ANIMATION_CONTROL = {
-  play: 'play',
-  pause: 'pause',
-  stop: 'stop',
-  loading: 'loading'
-};
+import propTypes, {ANIMATION_CONTROL} from './prop-types';
 
 const isIE11 = () => {
   if (typeof window === 'undefined') return;
@@ -179,22 +172,6 @@ const LottieWrapper = props => {
   );
 };
 
-LottieWrapper.propTypes = {
-  'aria-label': PropTypes.string.isRequired,
-  'data-name': PropTypes.string,
-  animationSrc: PropTypes.string.isRequired,
-  loop: PropTypes.bool,
-  rendererSettings: PropTypes.shape({
-    hideOnTransparent: PropTypes.bool,
-    className: PropTypes.string
-  }),
-  height: PropTypes.number,
-  width: PropTypes.number,
-  className: PropTypes.string,
-  ie11ImageBackup: PropTypes.string.isRequired,
-  backupImageClassName: PropTypes.string,
-  autoplay: PropTypes.bool,
-  animationControl: PropTypes.oneOf(keys(ANIMATION_CONTROL))
-};
+LottieWrapper.propTypes = propTypes;
 
 export default LottieWrapper;
