+++ packages/@coorpacademy-components/src/atom/review-background/prop-types.js
@@ -1,0 +1,3 @@
+import PropTypes from 'prop-types';
+
+export default {'aria-label': PropTypes.string};
