+++ packages/@coorpacademy-components/src/atom/review-header-step-item/prop-types.js
@@ -1,0 +1,7 @@
+import PropTypes from 'prop-types';
+
+export default {
+  icon: PropTypes.string,
+  current: PropTypes.bool,
+  value: PropTypes.string
+};
