+++ packages/@coorpacademy-components/src/atom/review-presentation/index.js
@@ -10,6 +10,7 @@ import {
   NovaLineSelectionCursorsCursorArrowTarget as TargetIcon
 } from '@coorpacademy/nova-icons';
 import style from './style.css';
+import propTypes from './prop-types';
 
 const ICONS = {
   skills: TargetIcon,
@@ -74,25 +75,10 @@ const ReviewPresentation = props => {
   );
 };
 
-const levelItem = PropTypes.shape({
-  text: PropTypes.string,
-  tooltipText: PropTypes.string
-});
-
 ToolTip.propTypes = {
   tooltipText: PropTypes.string
 };
 
-ReviewPresentation.propTypes = {
-  'aria-label': PropTypes.string,
-  reviewTitle: PropTypes.string,
-  reviewText: PropTypes.string,
-  labelsList: PropTypes.shape({
-    skills: levelItem,
-    questions: levelItem,
-    lifes: levelItem,
-    allright: levelItem
-  })
-};
+ReviewPresentation.propTypes = propTypes;
 
 export default ReviewPresentation;
