+++ packages/@coorpacademy-components/src/molecule/review-header-steps/prop-types.js
@@ -1,0 +1,7 @@
+import PropTypes from 'prop-types';
+
+import ReviewHeaderStepItemPropTypes from '../../atom/review-header-step-item/prop-types';
+
+export default {
+  steps: PropTypes.arrayOf(PropTypes.shape(ReviewHeaderStepItemPropTypes))
+};
