+++ packages/@coorpacademy-components/src/molecule/skill-card/prop-types.js
@@ -1,0 +1,13 @@
+import PropTypes from 'prop-types';
+
+export default {
+  'aria-label': PropTypes.string.isRequired,
+  skillTitle: PropTypes.string,
+  skillAriaLabel: PropTypes.string,
+  buttonLabel: PropTypes.string,
+  buttonAriaLabel: PropTypes.string,
+  reviseLabel: PropTypes.string,
+  reviseAriaLabel: PropTypes.string,
+  isCustom: PropTypes.bool,
+  onClick: PropTypes.func
+};
