+++ packages/@coorpacademy-components/src/organism/review-skills/prop-types.js
@@ -1,0 +1,21 @@
+import PropTypes from 'prop-types';
+import SkillCardPropTypes from '../../molecule/skill-card/prop-types';
+
+export const ListSkillsPropTypes = {
+  listSkills: PropTypes.arrayOf(SkillCardPropTypes)
+};
+
+export const NoSkillsPropTypes = {
+  titleNoSkills: PropTypes.string,
+  textNoSkills: PropTypes.string,
+  iconSkillAriaLabel: PropTypes.string
+};
+
+export default {
+  'aria-label': PropTypes.string,
+  title: PropTypes.string.isRequired,
+  isLoading: PropTypes.bool,
+  isLoadingAriaLabel: PropTypes.string,
+  ...ListSkillsPropTypes,
+  ...NoSkillsPropTypes
+};
